requirement.txt
streamlit: 1.51.0
authlib: 1.6.5
matplotlib: 3.10.6
pandas: 2.3.2
plotly: 6.5.0
pymongo: 4.15.4
python-dotenv>=1.2.1